\name{BALAGURANATHAN}
\alias{BALAGURANATHAN}
\docType{data}
\title{
 BALAGURUNATHAN Study Pancreatic Cancer Dataset
}
\description{
An expresion set object containing pancreatic cancer data
}
\format{
    SummarizedExperiment object
}
\details{
    more details can be found in the experimentData section of the object
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE11838
}
\references{
Balagurunathan et al, Mol Cancer Ther 2008
}
\examples{
BALAGURANATHANEset = loadPancreasEsets()$esets$BALAGURANATHAN
 experimentData(BALAGURANATHANEset)
}
\keyword{datasets}
