\name{ICGCMICRO}
\alias{ICGCMICRO}
\docType{data}
\title{
 ICGCMICRO Study Pancreatic Cancer Dataset
}
\description{
An expresion set object containing pancreatic cancer data
}
\format{
    eSet object
}
\details{
    more details can be found in the experimentData section of the object
}
\source{
http://icgc.org/icgc/cgp/68/304/798
}
\references{
Nones et al, Int. J. Cancer, 2014
}
\examples{
ICGCMICROEset = loadPancreasEsets()$esets$ICGCMICRO
}
\keyword{datasets}
