\name{PCSI}
\alias{PCSI}
\docType{data}
\title{
 PCSI Study Pancreatic Cancer Dataset
}
\description{
An expresion set object containing pancreatic cancer data
}
\format{
    eSet object
}
\details{
    more details can be found in the experimentData section of the object
}
\source{
URL unavailable (private dataset)
}
\references{
Notta et al, Nature 2016 
}
\examples{
PCSIEset = loadPancreasEsets()$esets$PCSI
 experimentData(PCSIEset)

}
\keyword{datasets}
