\name{BALAGURUNATHAN}
\alias{BALAGURUNATHAN}
\docType{data}
\title{
 BALAGURUNATHAN Study Pancreatic Cancer Dataset
}
\description{
An expresion set object containing pancreatic cancer data
}
\format{
    eSet object
}
\details{
    more details can be found in the experimentData section of the object
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE11838
}
\references{
Balagurunathan et al, Mol Cancer Ther 2008
}
\examples{
BALAGURUNATHEset = loadPancreasEsets()$esets$BALAGURUNATH
 experimentData(BALAGURUNATHEset)
}
\keyword{datasets}
